пофиксена логика метода удаления сотрудника, если он глава отдела в class EmployeeServiceImpl

+++ src/main/java/ru/technoserv/services/impl/EmployeeServiceImpl.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.technoserv.dao.*;
+import ru.technoserv.domain.Department;
 import ru.technoserv.domain.Employee;
 import ru.technoserv.domain.EmployeeHistory;
 import ru.technoserv.exceptions.EmployeeTheHeadOfDepartment;
@@ -54,6 +55,13 @@ public class EmployeeServiceImpl implements EmployeeService {
     @Override
     public void removeEmployee(int id) {
         logger.info("Удаляем сотрудника");
+        Department empDept = employeeDao.read(id).getDepartment();
+        int empDeptHeadID = empDept.getDeptHeadId();
+
+        if (empDeptHeadID == id) {
+            logger.info("Попытка удалить начальника отдела. Действие невозможно");
+            throw new EmployeeTheHeadOfDepartment(id);
+        }
         employeeDao.delete(id);
     }
